refactor: merges the cell of popular movies and similar movies into a single cell
