+++ PopMovies/PopMovies/Sources/Common/Components/MovieCell.swift
@@ -1,0 +1,75 @@
+//
+//  MovieCell.swift
+//  PopMovies
+//
+//  Created by Angela Alves on 03/01/23.
+//
+
+import Foundation
+import Kingfisher
+import TinyConstraints
+import UIKit
+
+class MovieCell: UICollectionViewCell {
+
+    static let identifier = "movieCell"
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configureSubViews()
+        configureConstraints()
+        backgroundColor = Theme.currentTheme.color.cellBackgroundColor.rawValue
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    private let imageView = UIImageView() .. {
+        $0.contentMode = .scaleAspectFill
+        $0.clipsToBounds = true
+        $0.layer.cornerRadius = 8.0
+    }
+
+    private let title = UILabel() .. {
+        $0.lineBreakMode = .byWordWrapping
+        $0.numberOfLines = 0
+        $0.textAlignment = .center
+        $0.font = UIFont.systemFont(ofSize: 16, weight: .bold)
+        $0.textColor = Theme.currentTheme.color.textColor.rawValue
+        $0.height(300)
+        $0.width(100)
+    }
+
+    func configureSubViews() {
+        addSubview(imageView)
+        addSubview(title)
+        layer.shadowColor = UIColor.black.cgColor
+        layer.shadowOpacity = 0.5
+        layer.cornerRadius = 8.0
+    }
+
+    func configureConstraints() {
+        imageView.edgesToSuperview()
+        title.centerInSuperview()
+    }
+
+    func setup(for movie: MovieItem) {
+        let url = URL(string: "https://image.tmdb.org/t/p/w500\(movie.posterPath ?? "")")
+        imageView.kf.indicatorType = .activity
+        imageView.kf.setImage(
+            with: url
+        ) { [weak self] result in
+            switch result {
+                case .success:
+                    self?.title.isHidden = true
+                    return
+                case .failure:
+                    self?.title.isHidden = false
+                    self?.title.text = movie.title
+                    return
+            }
+        }
+    }
+}
